* macosx/tkMacOSXWm.c (WmAttributesCmd, WmIconbitmapCmd): add support 	* unix/tkUnixSend.c (Tk_GetUserInactiveTime):             for weakly 	importing symbols not available on OSX 10.2 or 10.3, enables binaires 	built on later OSX versions to run on earlier ones. 	* macosx/Wish.xcodeproj/project.pbxproj: enable weak-linking; turn on 	                                         extra warnings. 	* macosx/README: document how to enable weak-linking; cleanup. 	* unix/configure.in: add check on Darwin-X11 for ld support of -weak-l 	* unix/tcl.m4:       flag and weak-link libXss if possible as it is not 	available before OSX 10.4; enforce requirement of OSX 10.2 for TkAqua; 	move Darwin specific checks & defines that are only relevant to the tcl 	build out of tcl.m4; restrict framework option to Darwin; cleanup 	quoting and help messages. 	* unix/configure: autoconf-2.59 	* unix/tkConfig.h.in: autoheader-2.59
	* macosx/GNUmakefile: enable xft for TkX11 build.
	* macosx/tkMacOSXFont.c (TkMacOSXQuarzStartDraw, TkMacOSXQuarzEndDraw):
	verify validity of context returned from QDBeginCGContext() before use.
	* macosx/tkMacOSXKeyEvent.c: ifdef out diagnostic messages to stderr.

	* macosx/tkMacOSXEvent.h:      standardize MAC_OS_X_VERSION_MAX_ALLOWED
	* macosx/tkMacOSXMenu.c:       checks per QA1316, ensure define can be
	* macosx/tkMacOSXMenubutton.c: overriden on command line (from default
	* macosx/tkMacOSXMenus.c:      of current OS version).
	* macosx/tkMacOSXMouseEvent.c:
	* macosx/tkMacOSXWm.c:

	* macosx/tkMacOSXFont.c (TkpMeasureCharsInContext): fix
	signed-with-unsigned comparison and other warnings from gcc4 -Wextra.

+++ unix/tkUnixSend.c
@@ -11,7 +11,7 @@
  * See the file "license.terms" for information on usage and redistribution of
  * this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tkUnixSend.c,v 1.14 2005/11/27 02:36:16 das Exp $
+ * RCS: @(#) $Id: tkUnixSend.c,v 1.15 2006/07/20 06:25:20 das Exp $
  */
 
 #include "tkPort.h"
@@ -1926,7 +1926,7 @@ TkpTestsendCmd(
 		    &bytesAfter, (unsigned char **) &property);
 	    if ((result == Success) && (actualType != None)
 		    && (actualFormat == 8) && (actualType == XA_STRING)) {
-		for (p = property; (p-property) < length; p++) {
+		for (p = property; (unsigned long)(p-property) < length; p++) {
 		    if (*p == 0) {
 			*p = '\n';
 		    }
